requirements.txt
streamlit>=1.35.0
tensorflow>=2.15.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0